+++ Source/Animal_Inheritance/Patches/Patch_RJW_BestialityPregnancyUtility.cs
@@ -30,7 +30,6 @@ namespace RJW_BGS
                         baby.genes = new Pawn_GeneTracker(baby);
                     }
 
-
                     //Remove the hair and skin genes pawns always start with, should get correct ones from human parent anyway.
                     for (int i = baby.genes.Endogenes.Count - 1; i >= 0; i--)
                     {
@@ -47,6 +46,8 @@ namespace RJW_BGS
 
                     //reason = "AbilityDisabledNoFertilinGene".Translate(this.parent.pawn);
                     baby.genes.xenotypeName = "rjw_genes_bestiality_hybrid_name".Translate();
+
+                    RJW_Genes.ModLog.Debug($"Birth of a Bestiality-Child - Added {humangenes.Count} human-genes and {beastgenes.Count} animal-genes.");
                 }
             }
         }