+++ Source/Animal_Inheritance/Settings/RJW_BGSSettings.cs
@@ -32,8 +32,6 @@ namespace RJW_BGS
             listing_Standard.Gap(5f);
             listing_Standard.CheckboxLabeled("rjw_genes_animal_inheritance_settings_ve_genetics_hybridization_key".Translate(), ref rjw_bgs_VE_genetics, "rjw_genes_animal_inheritance_settings_ve_genetics_hybridization_explanation".Translate(), 0f, 1f);
             listing_Standard.Gap(5f);
-            listing_Standard.CheckboxLabeled("rjw_genes_animal_inheritance_settings_detailed_debug_key".Translate(), ref rjw_bgs_detailed_debug, "rjw_genes_animal_inheritance_settings_detailed_debug_explanation".Translate(), 0f, 1f);
-            listing_Standard.End();
         }
 
         public override void ExposeData()
@@ -42,14 +40,12 @@ namespace RJW_BGS
             Scribe_Values.Look<bool>(ref RJW_BGSSettings.rjw_bgs_enabled, "rjw_bgs_enabled", RJW_BGSSettings.rjw_bgs_enabled, true);
             Scribe_Values.Look<float>(ref RJW_BGSSettings.rjw_bgs_global_gene_chance, "rjw_bgs_global_gene_chance", RJW_BGSSettings.rjw_bgs_global_gene_chance, true);
             Scribe_Values.Look<bool>(ref RJW_BGSSettings.rjw_bgs_animal_genes_as_xenogenes, "rjw_bgs_animal_genes_as_xenogenes", RJW_BGSSettings.rjw_bgs_animal_genes_as_xenogenes, true);
-            Scribe_Values.Look<bool>(ref RJW_BGSSettings.rjw_bgs_detailed_debug, "rjw_bgs_detailed_debug", RJW_BGSSettings.rjw_bgs_detailed_debug, true);
             Scribe_Values.Look<bool>(ref RJW_BGSSettings.rjw_bgs_VE_genetics, "rjw_bgs_VE_genetics", RJW_BGSSettings.rjw_bgs_VE_genetics, true);
         }
 
         public static float rjw_bgs_global_gene_chance = 1f;
         public static bool rjw_bgs_enabled = true;
         public static bool rjw_bgs_animal_genes_as_xenogenes = false;
-        public static bool rjw_bgs_detailed_debug = false;
         public static bool rjw_bgs_VE_genetics = true;
     }
 }